+++ src/extension.ts
@@ -10,10 +10,12 @@ import {registerCppToolsConfigurationProvider} from './ExternalApi/CppToolsApi';
 import {registerTaskProviders as registerBuildTaskProviders} from './BrAsBuildTaskProvider';
 import {registerTaskProviders as registerTransferTaskProviders} from './BrAsTransferTaskProvider';
 import {registerApiTests} from './Tools/ApiTests';
-import {registerProjectWorkspace} from './BRAsProjectWorkspace';
+import {getWorkspaceProjects, registerProjectWorkspace} from './BRAsProjectWorkspace';
 import { notifications } from './UI/Notifications';
 import { extensionState } from './BrExtensionState';
 import { extensionConfiguration } from './BRConfiguration';
+import { getAvailableAutomationStudioVersions, getAvailablePviVersions } from './BREnvironment';
+import { statusBar } from './UI/StatusBar';
 
 
 // Activation event
@@ -24,7 +26,7 @@ export async function activate(context: vscode.ExtensionContext) {
 		prettyPrintAdditionalData: extensionConfiguration.logging.prettyPrintAdditionalData
 	};
 	logger.info('Start activation of B&R Automation Tools extension');
-	//
+	// Initialize modules
 	extensionState.initialize(context);
 	notifications.initialize(context);
 	notifications.newVersionMessage();
@@ -32,8 +34,22 @@ export async function activate(context: vscode.ExtensionContext) {
 	registerCommands(context);
 	registerBuildTaskProviders(context);
 	registerTransferTaskProviders(context);
+	// get promises for long running activation events and add to status bar
+	const waitAsVersion = getAvailableAutomationStudioVersions();
+	statusBar.addBusyItem(waitAsVersion, 'Searching for installed AS versions');
+	const waitPviVersions = getAvailablePviVersions();
+	statusBar.addBusyItem(waitPviVersions, 'Searching for installed PVI versions');
+	const waitWorkspaceProjects = getWorkspaceProjects();
+	statusBar.addBusyItem(waitWorkspaceProjects, 'Parsing AS projects in workspace');
+	// TODO do we need to await these? Will probably be remove after architectural changes #5
 	await registerCppToolsConfigurationProvider(context);
 	await registerProjectWorkspace(context);
+	// Show activation message and log entry when all is done
+	await Promise.all([
+		waitAsVersion,
+		waitPviVersions,
+		waitWorkspaceProjects,
+	]);
 	logger.info('Activation of B&R Automation Tools extension finished');
 	notifications.activationMessage();
 }